Avoid spraying inline always everywhere
The inline always attribute should be avoided unless thorough
benchmarking has shown that it is helpful in a specific place. The
original thought was that adding it would improve vectorization, but it
is likely that the compiler is able to make correct decisions on its own
on where to inline.

Removing these attributes seemed to show a minor (0-2%) performance
improvement.

Additionally, adding the `#[inline]` attribute on a non-public function
does nothing, as this hint allows functions to be inlined across crates,
which means it is only useful on public functions.